Fix issues about visibility changes
1) The animation could be played although it is not visible.
onVisibilityChanged callback will be called even when
the visibility of an ancestor of the view has changed.
So we need to check the visibility with isShown(),
not with visibility == View.VISIBLE in onVisibilityChanged callback.

2) playAnimation() and resumeAnimation() could be called while not visible.
playAnimation() and resumeAnimation() should be blocked
if the view is not visible, and the animation should be
resumed when the view goes visible.
So, we need to set wasAnimatingWhenNotShown=true
in playAnimation() and resumeAnimation() when the view is not visible.

3) pauseAnimation() and cancelAnimation() were not worked when the activity is in background.
The animation could be resumed by onVisibilityChanged() callback.
We need to reset wasAnimatingWhenNotShown=false in pauseAnimation() and cancelAnimation().