Fixed render mode (#1377)
This change (https://github.com/airbnb/lottie-android/pull/1190) was supposed to revert to software rendering when the device can't allocate a hardware buffer large enough for the animation. However, this would also trigger before a composition loads and the view was set to wrap_content. The width and heigh would be 0 which would cause Android to noop the drawing cache. However, the previous logic here would think that the drawing cache failed to allocate because it was too large.

I also added some systrace markers and added a default animation view to the issue repro module.

Fixes #1374