Downgrade espresso and AWS (#1885)
#1874 broke instrumentation tests:

Espresso broke with this erro. It can probably be fixed by upgrading androidx test core but then there is a dependency conflict error which can probably be fixed by upgrading androidx.fragment:fragment-testing but there isn't a version of it that depends on test core 1.4.0.
The upgraded AWS SDK returns multiple status from the listener which caused the await continuation to be already resumed. I don't have time to debug the full behavior change and the old one was working fine.