Add a benchmark that can generate Baseline Profiles for Lottie.
* Add `LottieBaselineProfiles.kt` that can generate Baseline Profiles.
* Add `LottieBaselineBenchmark.kt` to benchmark the effect of Baseline Profiles.
* Add `baseline-prof.txt` files for `lottie` and `lottie/compose`.

```
LottieBaselineBenchmark_benchmarkNoCompilation
  timeToInitialDisplayMs   min   507.4,   median   507.4,   max   507.4
          frameCpuTimeMs   P50     16.8,   P90     31.2,   P95     52.6,   P99    122.4
    Traces: Iteration 0

LottieBaselineBenchmark_benchmarkBaselineProfiles
  timeToInitialDisplayMs   min 417.0,   median 417.0,   max 417.0
          frameCpuTimeMs   P50   16.3,   P90   21.5,   P95   22.6,   P99   44.7
    Traces: Iteration 0
```

To regenerate baseline profiles, just re-run the test defined in
`LottieBaselineProfiles.kt`. This should produce an output file with
the optimal baseline profile for the app defined in `sample-compose`.

To extract the relevant portions, use the tool defined [here](https://cs.android.com/androidx/platform/frameworks/support/+/androidx-main:development/splitBaselineProfiles).
This tool takes 2 command line arguments:

`profilePath` - This is the path to the baseline profile generated by the test.
`checkoutPath` - This is the path to the folder that contains `settings.gradle` (the module root).

This tool generates an output file, that splits the baseline profile
based on the module layout defined in the checkout path.