[Compose] Add LottieCompositionSpec.ContentProvider (#1982)
This adds a new LottieCompositionSpec option to load an animation from a content provider URI. This wasn't needed pre-Compose because you could get the input stream and parse the animation using that. However, Lottie Compose doesn't provide a direct API to parse an animation from an input stream. It doesn't expose that because a closeable input stream isn't a great candidate to put in a data class and makes LottieCompositionSpec stateful and closeable.

I also noticed some inconsistency in the way cache keys were handled for other specs so I fixed those as well.

Fixes #1978