Clear onVisibleAction once the action is taken (#2066)
The root cause of #2053 is that once the fragment was paused and resumed, the onResumeAction was set to RESUME and never cleared. The compass view in the linked project triggered a relayout on every frame which caused the visibility action run over and over again. The animation itself happens to have an empty first frame so it appeared as if the animation disappeared.

Fixes #2053