Add the LottieConfig.Builder.setEnableNetworkCache() method to completely disable internal network cache if needed (#2158)
**Rationale**
We'd like to implement more sophisticated caching strategy for animations fetched from network. This could be implemented on the network fetcher level e.g. using the OkHttp's cache or the "stale-while-revalidate" strategy.
For this to work the internal Lottie network cache has to be completely disabled.
The option to do so is implemented in this PR.

**Implementation**
By default the internal cache is enabled to maintain backward compatibility. 
We've added the `LottieConfig.Builder.setEnableNetworkCache()` method to override the default behavior.