Overhaul text layout (#2162)
This PR completely rewrites the text layout algorithm. It should be more accurate in most cases now. Center and right justifications as well as alignment of paragraph text and text wrapping should all be much more accurate.

Note that there may be an outstanding issue with the y position of paragraph text which affects #1758 but I have been unable to reproduce it with an AEP file for debugging.

Fixes #1479
Fixes #1758 
Fixes #397 
Likely fixes #2135