Update point text y offset (#2243)
After several hours and a bunch of test files, I've been able to get the y position of point text to be accurate in most cases and closer in the rest. I can't quite get a hold of the right algorithm to get it to be 100% correct all the time but this is an improvement.

If any external contributors wants to continue iterating on this, feel free to. Just add more animations to the snapshot-test module if you do.

Mostly fixes #2201