Fixes flickering when the same URL is set to the LottieAnimationView
When the same URL is set in a LottieAnimationView multiple times, the view
continuously clears and resets the composition from the cache resulting in
flickering. This change fixes the flickering by adding the currently loaded
URL as part of the view state. If the same URL is set multiple times, the
LottieAnimationView now knows not to reload the composition again and again
and thus avoid flickering.