Fixes flickering when the same resource is set to the LottieAnimationView
When the same resource (url, json, etc) is set in a LottieAnimationView
multiple times, the view continuously clears and resets the composition
resulting in flickering. This change fixes flickering by not eagerly
clearing the LottieAnimationView, but instead keeping the old animation
displayed until the new composition is available and can be rendered.
The view is cleared only when there is a failure loading the new resource.