Add support for Multiply blend mode
BlendModeCompat is designed to use either a BlendMode (added in Android
Q) or PorterDuff.Mode (always available).

Our support for Lottie blend modes did not include Multiply due to a
slightly different formula between the PorterDuff and BlendMode
variants.

However, we did include support for Screen, which suffers from a similar
behavior. Therefore, we are not breaking any consistency by including
Multiply too, and including it provides benefits in terms of more
complete support, as Multiply is a foundational blend mode.