Adjust drop shadow distance and softness
This applies a constant scale factor to the distance and softness (aka blur or radius) properties on drop shadow effects so that the end result more closely matches what is shown in After Effects and on other platforms.

See https://github.com/airbnb/lottie-ios/pull/2175 for similar changes made to the iOS Lottie implementation.