Apply transformation for the stroke on the canvas so the skew is also… (#2531)
I noticed that the Android implementation had trouble with applying a skew on a stroke, so this attempts to fix that issue.

The issue was that since the transformation was applied to the path, it didn’t take any properties from the paint into consideration. As a result, the skew was not applied to the width of the line. By applying the transformation on the canvas, this issue is resolved.

Example Lottie which wasn’t rendered correctly (should have skewed lines):
[Skewed-Stroke.json](https://github.com/user-attachments/files/16495718/Skewed-Stroke.json)

Co-authored-by: Mirko <m.volkers@wearetriple.com>