Drop shadow overhaul: improve correctness and perf
This commit introduces a large change to how drop shadows are rendered,
introducing an `applyShadowsToLayers` flag which, by analogy to
`applyOpacitiesToLayers`, allows layers to be treated as a whole for the
purposes of drop shadows, improving the accuracy and bringing
lottie-android in line with other renderers (lottie-web and lottie-ios).

Several different codepaths for different hardware/software combinations
are introduced to ensure the fastest rendering available, even on legacy
devices.

The calculation of shadow direction with respect to transforms is
improved so that the output matches lottie-web and lottie-ios.

Image layers now cast shadows correctly thanks to a workaround to
device-specific issues when combining `Paint.setShadowLayer()` and
bitmap rendering.

Even in non-`applyShadowsToLayers` mode, correctness is improved by
allowing the shadow-to-be-applied to propagate in a similar way as
alpha. This allows some amount of visual fidelity to be recovered for
animations or environments where enabling `applyShadowsToLayers` is not
possible.

A number of issues that caused incorrect rendering in some other cases
have been fixed.