Apply DP scaling to canvas in software rendering
When rendering a view, the Android graphics stack, by convention, gives
it a canvas where the transformation to put the view in the right
position is applied. It also allows the view to draw with its intrinsic
width and height in pixels, and the canvas has an inverse scaling
transform to draw to the final surface with the right size.

Thus, the transform matrices we work with during hardware rendering
exclude this inverse scaling.

In software rendering mode, we bake this scaling into the root matrix,
which causes different values to appear in software vs hardware
rendering, which:
* Has an impact on the final image due to e.g. integer rounding
* Makes debugging HW/SW differences trickier, as values are different.

To make the modes closer, "unbake" the pre-existing scale of the canvas
from the initial root matrix.

To mitigate risk, this is only done for scale, and not for the full
matrix itself, as there is code which might interact with other types of
transforms in unexpected ways. (Whereas scaling is a "safer" one.) This
can be done in the future to achieve 100% equal values between the two
modes.