Fix image bounds calculation
`ImageLayer.getBounds()` works with a parent-to-world matrix called
`parentMatrix`, but `ImageLayer.drawLayer()` gets called with
layer-to-world as `parentMatrix`, despite the name. (See call to
`drawLayer() `in `BaseLayer.draw()`).

This caused `getBounds()` to apply the layer's own transform effectively
two times and the returned bounds to be incorrect.

With non-shadowed layers, this was masked away by the fact that
`Canvas.saveLayer()` appears to ignore the provided bounds Rect, and
allows anything drawn anywhere on the canvas to appear after a
corresponding `canvas.restore()`.

On top of this, we erroneously multiplied the returned `bounds` by the
density again, which meant that the final bitmap was "overprovisioned"
by a factor of 2-3 (depending on the device's pixel density). In most
cases, and combined with an erroneous order of operations when applying
matrices to the original and off-screen canvases, this was enough to
hide the problem even on shadowed layers.

Fix by:
* Not relying on `getBounds()`, as we don't have a correct
  `parentMatrix` to feed it.
* Correcting the order of operations, so that we respect the convention
  of not applying any matrix to the canvas before calling
  `OffscreenLayer.start()`, which simplifies the code and allows us to
  use the same pattern as other invocations of `OffscreenLayer`.
* Getting rid of the unnecessary scale by `density` and then
  normalization back, borne out of a misunderstanding of why `density`
  is a factor in this calculation. (It effectively gets applied to *all*
  values somewhere in the pipeline, and has nothing to do with images
  per se.)

Drive-by fix `getBounds()` itself to respect
`maintainOriginalImageBounds`.