Much faster CPU blur
Based on the old box blur algorithm with a running sum, write a new one
that is more readable and more performant, due mostly to avoiding the
use of Bitmap.getPixels() and Bitmap.setPixels().

The blur is still, well, boxy, which we will probably be able to fix by
stacking two invocations of it, or trying some trickery to approximate
the Guassian kernel better but still keeping the linear time complexity.